Separate Navbar component from Layout file
